# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Exchange server API 1.0 swagger doc",
        "title": "Exchange server API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "{{.BasePath}}",
    "paths": {
        "/order_books": {
            "get": {
                "description": "get your order book list",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "order book list",
                "parameters": [
                    {
                        "type": "string",
                        "description": "eg BTC_USD",
                        "name": "symbol",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.OrderBook"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "create an order book",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "new an order book",
                "parameters": [
                    {
                        "description": "order book model",
                        "name": "req",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/models.OrderBook"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/models.OrderBook"
                        }
                    }
                }
            }
        },
        "/order_books/{id}": {
            "put": {
                "description": "update an exist order book",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "update an order book",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "eg 1",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "order book model",
                        "name": "req",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/models.OrderBook"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/models.OrderBook"
                        }
                    }
                }
            },
            "delete": {
                "description": "cancel an exist order book",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "cancel an order book",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "eg 1",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/utils.APIRes"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "models.Currency": {
            "type": "object",
            "properties": {
                "deposit_fee": {
                    "type": "number"
                },
                "symbol": {
                    "type": "string"
                },
                "withdraw_fee": {
                    "type": "number"
                }
            }
        },
        "models.Fund": {
            "type": "object",
            "properties": {
                "leftCurrency": {
                    "type": "object",
                    "$ref": "#/definitions/models.Currency"
                },
                "leftCurrencyID": {
                    "type": "integer"
                },
                "limit_rate": {
                    "type": "number"
                },
                "market_rate": {
                    "type": "number"
                },
                "name": {
                    "type": "string"
                },
                "rightCurrency": {
                    "type": "object",
                    "$ref": "#/definitions/models.Currency"
                },
                "rightCurrencyID": {
                    "type": "integer"
                },
                "symbol": {
                    "description": "eg btc_usd eth_usd",
                    "type": "string"
                }
            }
        },
        "models.OrderBook": {
            "type": "object",
            "properties": {
                "fund": {
                    "type": "object",
                    "$ref": "#/definitions/models.Fund"
                },
                "fund_id": {
                    "type": "integer"
                },
                "order_type": {
                    "description": "market or limit",
                    "type": "string"
                },
                "price": {
                    "type": "number"
                },
                "side": {
                    "description": "Sell Buy",
                    "type": "string"
                },
                "symbol": {
                    "type": "string"
                },
                "user": {
                    "type": "object",
                    "$ref": "#/definitions/models.User"
                },
                "user_id": {
                    "type": "integer"
                },
                "volume": {
                    "type": "number"
                }
            }
        },
        "models.User": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "role": {
                    "type": "string"
                }
            }
        },
        "utils.APIRes": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "message": {
                    "type": "string"
                }
            }
        }
    }
}